Resolvers codesplit
