+++ package.json
@@ -35,6 +35,7 @@
     "@graphql-codegen/typescript": "3.0.0",
     "@graphql-codegen/typescript-resolvers": "3.0.0",
     "@types/lodash": "4.14.191",
+    "@types/node": "18.11.18",
     "@yandex-cloud/function-types": "2.1.0",
     "@yandex-cloud/nodejs-sdk": "2.3.2",
     "dotenv": "16.0.3",
